[PF-48] - Builder integration changes
- merge canvas configuration service with canvas service
