+++ projects/example-app/src/app/app.component.html
@@ -45,7 +45,7 @@
             <mat-button-toggle mat-button (click)="openDialog()">
                 <mat-icon>info</mat-icon>
             </mat-button-toggle>
-            <mat-button-toggle mat-button (click)="petriflowConfigService.gridOnOff()">
+            <mat-button-toggle mat-button (click)="petriflowCanvasService.gridOnOff()">
                 <mat-icon>clear</mat-icon>
             </mat-button-toggle>
             <mat-button-toggle mat-button id="reset-zoom-button" (click)="resetPanZoom()">
@@ -75,5 +75,5 @@
 </mat-toolbar>
 <div class="outer">
     <petriflow-svg-canvas class="canvas" [panzoomConfiguration]="panzoomConfiguration"
-                          [gridConfiguration]="petriflowConfigService.gridConfiguration"></petriflow-svg-canvas>
+                          [gridConfiguration]="petriflowCanvasService.gridConfiguration"></petriflow-svg-canvas>
 </div>
