+++ projects/example-app/src/app/app.component.ts
@@ -10,9 +10,6 @@ import {PetriflowInfoDialogComponent} from './petriflow-info-dialog/petriflow-in
 import {Place, Transition} from "@netgrif/petri.svg";
 import {ControlPetriflowCanvasService} from "./petriflow-control-set/services/control-petriflow-canvas.service";
 import {MatSnackBar} from "@angular/material/snack-bar";
-import {
-    ControlPetriflowCanvasConfigurationService
-} from "./petriflow-control-set/services/control-petriflow-canvas-configuration.service";
 import {CanvasMode} from './petriflow-control-set/domain/canvas-mode';
 
 @Component({
@@ -20,8 +17,7 @@ import {CanvasMode} from './petriflow-control-set/domain/canvas-mode';
     templateUrl: './app.component.html',
     styleUrls: ['./app.component.scss'],
     providers: [
-        {provide: PetriflowCanvasService, useExisting: ControlPetriflowCanvasService},
-        {provide: PetriflowCanvasConfiguration, useExisting: ControlPetriflowCanvasConfigurationService}
+        {provide: PetriflowCanvasService, useExisting: ControlPetriflowCanvasService}
     ]
 })
 export class AppComponent implements AfterViewInit {
@@ -44,40 +40,39 @@ export class AppComponent implements AfterViewInit {
         minScale: this.minScale
     };
 
-    constructor(private _petriflowCanvasService: ControlPetriflowCanvasService,
-                public petriflowConfigService: ControlPetriflowCanvasConfigurationService,
+    constructor(public petriflowCanvasService: ControlPetriflowCanvasService,
                 public dialog: MatDialog,
                 private _snackBar: MatSnackBar) {
-        this._mode = petriflowConfigService.mode;
+        this._mode = petriflowCanvasService.mode;
     }
 
     ngAfterViewInit(): void {
-        if (!this._petriflowCanvasService.canvas) {
+        if (!this.petriflowCanvasService.canvas) {
             throw new Error('Petriflow SVG canvas does not exists!');
         }
-        this._petriflowCanvasService.canvas.svg.onclick = (e) => {
+        this.petriflowCanvasService.canvas.svg.onclick = (e) => {
             this.addTransition(e);
             this.addPlace(e);
         };
-        this.petriflowConfigService.addCanvasEvent(this._petriflowCanvasService.canvas.svg);
+        this.petriflowCanvasService.addCanvasEvent(this.petriflowCanvasService.canvas.svg);
     }
 
     private addTransition($event: MouseEvent): void {
-        if (this.petriflowConfigService.mode === CanvasMode.CREATE_TRANSITION) {
+        if (this.petriflowCanvasService.mode === CanvasMode.CREATE_TRANSITION) {
             const transition = new Transition(
                 `t${++PetriflowCanvasConfiguration.TRANSITION_ID_COUNTER}`,
                 `t${PetriflowCanvasConfiguration.TRANSITION_ID_COUNTER}`,
                 new DOMPoint($event.offsetX, $event.offsetY)
             );
             const petriflowTransition = new PetriflowTransition(transition, undefined);
-            this._petriflowCanvasService.createTransition(transition);
-            this._petriflowCanvasService.petriflowElementsCollection.transitions.push(petriflowTransition);
-            this.petriflowConfigService.addTransitionEvents(petriflowTransition);
+            this.petriflowCanvasService.createTransition(transition);
+            this.petriflowCanvasService.petriflowElementsCollection.transitions.push(petriflowTransition);
+            this.petriflowCanvasService.addTransitionEvents(petriflowTransition);
         }
     }
 
     private addPlace($event: MouseEvent) {
-        if (this.petriflowConfigService.mode === CanvasMode.CREATE_PLACE) {
+        if (this.petriflowCanvasService.mode === CanvasMode.CREATE_PLACE) {
             const place = new Place(
                 `p${++PetriflowCanvasConfiguration.PLACE_ID_COUNTER}`,
                 `p${PetriflowCanvasConfiguration.PLACE_ID_COUNTER}`,
@@ -85,9 +80,9 @@ export class AppComponent implements AfterViewInit {
                 new DOMPoint($event.offsetX, $event.offsetY)
             );
             const petriflowPlace = new PetriflowPlace(place);
-            this._petriflowCanvasService.createPlace(place);
-            this._petriflowCanvasService.petriflowElementsCollection.places.push(petriflowPlace);
-            this.petriflowConfigService.addPlaceEvents(petriflowPlace);
+            this.petriflowCanvasService.createPlace(place);
+            this.petriflowCanvasService.petriflowElementsCollection.places.push(petriflowPlace);
+            this.petriflowCanvasService.addPlaceEvents(petriflowPlace);
         }
     }
 
@@ -95,40 +90,40 @@ export class AppComponent implements AfterViewInit {
     onControlC($event: KeyboardEvent) {
         $event.preventDefault();
         this.openSnackBar('Canvas elements copied to clipboard');
-        this.petriflowConfigService.copyElements();
+        this.petriflowCanvasService.copyElementsToClipboard();
     }
 
     @HostListener('window:keydown.control.v', ['$event'])
     onControlV($event: KeyboardEvent) {
         $event.preventDefault();
-        this.petriflowConfigService.pasteElements();
+        this.petriflowCanvasService.pasteElements();
     }
 
     @HostListener('window:keydown.control.a', ['$event'])
     onControlA($event: KeyboardEvent) {
         $event.preventDefault();
         this.openSnackBar('Selected all petri-svg elements');
-        this._petriflowCanvasService.selectAll();
+        this.petriflowCanvasService.selectAll();
     }
 
     @HostListener('window:keydown.delete', ['$event'])
     onDelete() {
         this.openSnackBar('All selected petri-svg elements deleted');
-        this.petriflowConfigService.deleteSelectedElements();
+        this.petriflowCanvasService.deleteSelectedElements();
     }
 
     @HostListener('window:keydown.escape', ['$event'])
     onEscape() {
-        this._petriflowCanvasService.deselectAll();
-        this.petriflowConfigService.deleteClipboard();
-        this._petriflowCanvasService.panzoom?.reset();
+        this.petriflowCanvasService.deselectAll();
+        this.petriflowCanvasService.deleteClipboard();
+        this.petriflowCanvasService.panzoom?.reset();
     }
 
     @HostListener('window:keydown.+', ['$event'])
     onPlusButton() {
-        const newScale = !!this._petriflowCanvasService.panzoom?.getScale() ?
-            this._petriflowCanvasService.panzoom?.getScale() + this.stepSize : 1.2;
-        this._petriflowCanvasService.panzoom?.zoomToPoint(
+        const newScale = !!this.petriflowCanvasService.panzoom?.getScale() ?
+            this.petriflowCanvasService.panzoom?.getScale() + this.stepSize : 1.2;
+        this.petriflowCanvasService.panzoom?.zoomToPoint(
             newScale > this.maxScale ? this.maxScale : newScale,
             {
                 clientX: this._mouseEvent?.x ?? 0,
@@ -140,9 +135,9 @@ export class AppComponent implements AfterViewInit {
     // TODO: fix move on zoom
     @HostListener('window:keydown.-', ['$event'])
     onMinusButton() {
-        const newScale = !!this._petriflowCanvasService.panzoom?.getScale() ?
-            this._petriflowCanvasService.panzoom?.getScale() - this.stepSize : 1.2;
-        this._petriflowCanvasService.panzoom?.zoomToPoint(
+        const newScale = !!this.petriflowCanvasService.panzoom?.getScale() ?
+            this.petriflowCanvasService.panzoom?.getScale() - this.stepSize : 1.2;
+        this.petriflowCanvasService.panzoom?.zoomToPoint(
             newScale < this.minScale ? this.minScale : newScale,
             {
                 clientX: this._mouseEvent?.x ?? 0,
@@ -153,22 +148,22 @@ export class AppComponent implements AfterViewInit {
 
     @HostListener('window:keydown.ArrowUp', ['$event'])
     onUpButton() {
-        this._petriflowCanvasService.panzoom?.pan(0, PetriflowCanvasConfiguration.PANZOOM_MOVE, {relative: true});
+        this.petriflowCanvasService.panzoom?.pan(0, PetriflowCanvasConfiguration.PANZOOM_MOVE, {relative: true});
     }
 
     @HostListener('window:keydown.ArrowRight', ['$event'])
     onRightButton() {
-        this._petriflowCanvasService.panzoom?.pan(-PetriflowCanvasConfiguration.PANZOOM_MOVE, 0, {relative: true});
+        this.petriflowCanvasService.panzoom?.pan(-PetriflowCanvasConfiguration.PANZOOM_MOVE, 0, {relative: true});
     }
 
     @HostListener('window:keydown.ArrowDown', ['$event'])
     onDownButton() {
-        this._petriflowCanvasService.panzoom?.pan(0, -PetriflowCanvasConfiguration.PANZOOM_MOVE, {relative: true});
+        this.petriflowCanvasService.panzoom?.pan(0, -PetriflowCanvasConfiguration.PANZOOM_MOVE, {relative: true});
     }
 
     @HostListener('window:keydown.ArrowLeft', ['$event'])
     onLeftButton() {
-        this._petriflowCanvasService.panzoom?.pan(PetriflowCanvasConfiguration.PANZOOM_MOVE, 0, {relative: true});
+        this.petriflowCanvasService.panzoom?.pan(PetriflowCanvasConfiguration.PANZOOM_MOVE, 0, {relative: true});
     }
 
     @HostListener('mousemove', ['$event'])
@@ -181,8 +176,8 @@ export class AppComponent implements AfterViewInit {
     }
 
     changeCanvasMode(mode: CanvasMode) {
-        this.petriflowConfigService.disablePreviousArcMode();
-        this.petriflowConfigService.mode = mode;
+        this.petriflowCanvasService.disablePreviousArcMode();
+        this.petriflowCanvasService.mode = mode;
     }
 
     goToLink(url: string) {
@@ -194,7 +189,7 @@ export class AppComponent implements AfterViewInit {
     }
 
     resetPanZoom() {
-        this._petriflowCanvasService.panzoom?.reset();
+        this.petriflowCanvasService.panzoom?.reset();
     }
 
     openDialog() {
