--- projects/petriflow-svg/src/lib/services/public-api.ts
@@ -1,1 +1,0 @@
-export * from './petriflow-canvas-configuration.service';
